+++ org/gjt/sp/jedit/syntax/Token.java
@@ -27,7 +27,7 @@ import java.lang.reflect.Field;
  * A linked list of syntax tokens.
  *
  * @author Slava Pestov
- * @version $Id: Token.java,v 1.10 2003/03/16 20:55:48 spestov Exp $
+ * @version $Id: Token.java,v 1.11 2006/03/03 16:18:06 ezust Exp $
  */
 public class Token
 {
@@ -42,7 +42,7 @@ public class Token
 		try
 		{
 			Field f = Token.class.getField(value);
-			return (byte)f.getByte(null);
+			return f.getByte(null);
 		}
 		catch(Exception e)
 		{
